backend/Social_network/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.1
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==19.6.2
flake8==6.1.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.8.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.2
mccabe==0.7.0
packaging==23.1
phonenumbers==8.13.19
pika==1.3.2
Pillow==10.0.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.10
psycopg-binary==3.1.10
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
